Rename prepostcleaning to censoring.
