+++ xcp_d/interfaces/__init__.py
@@ -9,7 +9,7 @@ from xcp_d.interfaces import (
     execsummary,
     nilearn,
     plotting,
-    prepostcleaning,
+    censoring,
     report,
     report_core,
     restingstate,
@@ -24,7 +24,7 @@ __all__ = [
     "execsummary",
     "nilearn",
     "plotting",
-    "prepostcleaning",
+    "censoring",
     "report",
     "report_core",
     "restingstate",
