+++ xcp_d/workflows/postprocessing.py
@@ -12,7 +12,7 @@ from pkg_resources import resource_filename as pkgrf
 from xcp_d.interfaces.bids import DerivativesDataSink
 from xcp_d.interfaces.nilearn import DenoiseCifti, DenoiseNifti, Smooth
 from xcp_d.interfaces.plotting import CensoringPlot
-from xcp_d.interfaces.prepostcleaning import (
+from xcp_d.interfaces.censoring import (
     Censor,
     FlagMotionOutliers,
     RemoveDummyVolumes,
